third_party/blink/web_tests/virtual/user-activation-v2/fast/dom/Window/window-postmessage-user-gesture-expected.txt

Test that user gesture is available through postMessages to appropriate frames.
Click me

PASS Stack postMessages on self window and the user gesture is only consumed once
PASS Bounce postMessages between self and frame window and the user gesture is consumed only once in self window
PASS Bounce postMessages between self and frame window and the user gesture is consumed 0 times in frame
PASS successfullyParsed is true

TEST COMPLETE
